Re-organised SplineHelper module
