+++ SplineHelper/src/main/java/fr/radi3nt/spline/splines/dimensions/EncapsulatingSpline3D.java
@@ -30,4 +30,8 @@ public class EncapsulatingSpline3D implements Spline3D {
     private float velocityIndex(int index, float t) {
         return splines[index].velocity(t);
     }
+
+    public Spline[] getSplines() {
+        return splines;
+    }
 }
